+++ lib/reactive.dart
@@ -1,10 +1,13 @@
 export 'src/async/async_signal.dart' show AsyncSignal;
-export 'src/async/async_value.dart' show AsyncValue, AsyncData, AsyncLoading, AsyncError;
+export 'src/async/async_value.dart'
+    show AsyncValue, AsyncData, AsyncLoading, AsyncError;
 export 'src/computed.dart' show Computed;
 export 'src/consumer.dart' show batch;
-export 'src/effect.dart' show Effect, effect;
+export 'src/effect.dart' show Effect, effect, effectWithCleanup;
 export 'src/family.dart' show SignalFamily, AsyncSignalFamily;
-export 'src/observer.dart' show SignalObserver, signalObserver, ConsoleSignalObserver;
+export 'src/observer.dart'
+    show SignalObserver, signalObserver, ConsoleSignalObserver;
 export 'src/signal.dart' show Signal;
+export 'src/signal_subscription.dart' show SignalSubscription;
 export 'src/extensions.dart';
 export 'src/persistence.dart';